Fix: Clear old agreement data when viewing running negotiations
- Clear sessionComplete.value when loading a running negotiation to prevent showing agreement markers from previous runs
- Change default_time_limit type from int to float for consistency with other time limit fields

+++ negmas_app/models/settings.py
@@ -32,7 +32,7 @@ class NegotiationSettings:
 
     default_max_steps: int = 100
     default_step_delay_ms: int = 100
-    default_time_limit: int | None = None
+    default_time_limit: float | None = None
 
 
 @dataclass
